% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGeneric.R, R/unitSphere_innerProduct1.R,
%   R/unitSphere_polynomial.R
\name{domainCheckP}
\alias{domainCheckP}
\alias{domainCheckP,unitSphere_innerProduct1,matrix,list-method}
\alias{domainCheckP,unitSphere_polynomial,matrix,list-method}
\title{Check if node points are in the domain of a test function instance ("overload" of domainCheck with additional parameter)}
\usage{
domainCheckP(object, x, param)

\S4method{domainCheckP}{unitSphere_innerProduct1,matrix,list}(object, x, param)

\S4method{domainCheckP}{unitSphere_polynomial,matrix,list}(object, x, param)
}
\arguments{
\item{object}{Test function that gets evaluated}

\item{x}{Matrix where each row represents one evaluation point}

\item{param}{List of additional parameters}
}
\value{
Vector where each element (TRUE or FALSE) indicates if a point (row in the input matrix) is in the integration domain
}
\description{
domainCheckP is a generic function that allows to test if a collection of evaluation points are inside the integration domain associated to the test function instance or not.
This "overload" of domainCheck allows to pass a list of additional parameters.
}
\author{
Klaus Herrmann
}
