% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGeneric.R, R/Pn_lognormalDensity.R,
%   R/Pn_logtDensity.R, R/Rn_Gauss.R, R/Rn_floorNorm.R, R/Rn_normalDensity.R,
%   R/Rn_tDensity.R, R/standardSimplex_Dirichlet.R, R/standardSimplex_exp_sum.R,
%   R/unitBall_normGauss.R, R/unitBall_polynomial.R, R/unitCube_BFN4.R,
%   R/unitCube_cos2.R, R/unitCube_floor.R, R/unitCube_max.R,
%   R/unitSphere_innerProduct1.R, R/unitSphere_polynomial.R
\name{exactIntegral}
\alias{exactIntegral}
\alias{exactIntegral,Pn_lognormalDensity-method}
\alias{exactIntegral,Pn_logtDensity-method}
\alias{exactIntegral,Rn_Gauss-method}
\alias{exactIntegral,Rn_floorNorm-method}
\alias{exactIntegral,Rn_normalDensity-method}
\alias{exactIntegral,Rn_tDensity-method}
\alias{exactIntegral,standardSimplex_Dirichlet-method}
\alias{exactIntegral,standardSimplex_exp_sum-method}
\alias{exactIntegral,unitBall_normGauss-method}
\alias{exactIntegral,unitBall_polynomial-method}
\alias{exactIntegral,unitCube_BFN4-method}
\alias{exactIntegral,unitCube_cos2-method}
\alias{exactIntegral,unitCube_floor-method}
\alias{exactIntegral,unitCube_max-method}
\alias{exactIntegral,unitSphere_innerProduct1-method}
\alias{exactIntegral,unitSphere_polynomial-method}
\title{Get exact integral for test function instance}
\usage{
exactIntegral(object)

\S4method{exactIntegral}{Pn_lognormalDensity}(object)

\S4method{exactIntegral}{Pn_logtDensity}(object)

\S4method{exactIntegral}{Rn_Gauss}(object)

\S4method{exactIntegral}{Rn_floorNorm}(object)

\S4method{exactIntegral}{Rn_normalDensity}(object)

\S4method{exactIntegral}{Rn_tDensity}(object)

\S4method{exactIntegral}{standardSimplex_Dirichlet}(object)

\S4method{exactIntegral}{standardSimplex_exp_sum}(object)

\S4method{exactIntegral}{unitBall_normGauss}(object)

\S4method{exactIntegral}{unitBall_polynomial}(object)

\S4method{exactIntegral}{unitCube_BFN4}(object)

\S4method{exactIntegral}{unitCube_cos2}(object)

\S4method{exactIntegral}{unitCube_floor}(object)

\S4method{exactIntegral}{unitCube_max}(object)

\S4method{exactIntegral}{unitSphere_innerProduct1}(object)

\S4method{exactIntegral}{unitSphere_polynomial}(object)
}
\arguments{
\item{object}{The test function that gets evaluated}
}
\value{
Numeric value of the integral of the test function
}
\description{
exactIntegral is a generic function that allows to calculate the exact value of a test function instance over the associated integration domain.
}
\author{
Klaus Herrmann
}
