% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGeneric.R, R/Rn_Gauss.R, R/Rn_floorNorm.R,
%   R/standardSimplex_Dirichlet.R, R/standardSimplex_exp_sum.R,
%   R/unitBall_normGauss.R, R/unitBall_polynomial.R, R/unitCube_cos2.R,
%   R/unitCube_floor.R, R/unitSphere_innerProduct1.R, R/unitSphere_polynomial.R
\name{getIntegrationDomain}
\alias{getIntegrationDomain}
\alias{getIntegrationDomain,Rn_Gauss-method}
\alias{getIntegrationDomain,Rn_floorNorm-method}
\alias{getIntegrationDomain,standardSimplex_Dirichlet-method}
\alias{getIntegrationDomain,standardSimplex_exp_sum-method}
\alias{getIntegrationDomain,unitBall_normGauss-method}
\alias{getIntegrationDomain,unitBall_polynomial-method}
\alias{getIntegrationDomain,unitCube_cos2-method}
\alias{getIntegrationDomain,unitCube_floor-method}
\alias{getIntegrationDomain,unitSphere_innerProduct1-method}
\alias{getIntegrationDomain,unitSphere_polynomial-method}
\title{Get description of integration domain for test function instance}
\usage{
getIntegrationDomain(object)

\S4method{getIntegrationDomain}{Rn_Gauss}(object)

\S4method{getIntegrationDomain}{Rn_floorNorm}(object)

\S4method{getIntegrationDomain}{standardSimplex_Dirichlet}(object)

\S4method{getIntegrationDomain}{standardSimplex_exp_sum}(object)

\S4method{getIntegrationDomain}{unitBall_normGauss}(object)

\S4method{getIntegrationDomain}{unitBall_polynomial}(object)

\S4method{getIntegrationDomain}{unitCube_cos2}(object)

\S4method{getIntegrationDomain}{unitCube_floor}(object)

\S4method{getIntegrationDomain}{unitSphere_innerProduct1}(object)

\S4method{getIntegrationDomain}{unitSphere_polynomial}(object)
}
\arguments{
\item{object}{Test function for which the description is returned}
}
\value{
Description of the integration domain of the function
}
\description{
getIntegrationDomain is a generic function that returns a description of the integration domain associate to the test function instance.
}
\author{
Klaus Herrmann
}
