% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGeneric.R, R/Rn_Gauss.R, R/Rn_floorNorm.R,
%   R/standardSimplex_Dirichlet.R, R/standardSimplex_exp_sum.R,
%   R/unitBall_normGauss.R, R/unitBall_polynomial.R, R/unitCube_cos2.R,
%   R/unitCube_floor.R, R/unitSphere_innerProduct1.R, R/unitSphere_polynomial.R
\name{getReferences}
\alias{getReferences}
\alias{getReferences,Rn_Gauss-method}
\alias{getReferences,Rn_floorNorm-method}
\alias{getReferences,standardSimplex_Dirichlet-method}
\alias{getReferences,standardSimplex_exp_sum-method}
\alias{getReferences,unitBall_normGauss-method}
\alias{getReferences,unitBall_polynomial-method}
\alias{getReferences,unitCube_cos2-method}
\alias{getReferences,unitCube_floor-method}
\alias{getReferences,unitSphere_innerProduct1-method}
\alias{getReferences,unitSphere_polynomial-method}
\title{Get references for test function instance}
\usage{
getReferences(object)

\S4method{getReferences}{Rn_Gauss}(object)

\S4method{getReferences}{Rn_floorNorm}(object)

\S4method{getReferences}{standardSimplex_Dirichlet}(object)

\S4method{getReferences}{standardSimplex_exp_sum}(object)

\S4method{getReferences}{unitBall_normGauss}(object)

\S4method{getReferences}{unitBall_polynomial}(object)

\S4method{getReferences}{unitCube_cos2}(object)

\S4method{getReferences}{unitCube_floor}(object)

\S4method{getReferences}{unitSphere_innerProduct1}(object)

\S4method{getReferences}{unitSphere_polynomial}(object)
}
\arguments{
\item{object}{Test function for which the references are returned}
}
\value{
Vector with references for the specific function
}
\description{
getReferences is a generic function that returns a vector of references associated to the test function instance.
}
\author{
Klaus Herrmann
}
