% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multIntTestFunc.R
\name{multIntTestFunc}
\alias{multIntTestFunc}
\title{multIntTestFunc: A package to define test functions for multivariate numerical integration.}
\description{
The multIntTestFunc package provides multivariate test functions to test numerical integration routines.
}
\section{multIntTestFunc functions}{

The multIntTestFunc functions ...
}

