% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{quadratureProductRule}
\alias{quadratureProductRule}
\title{Product rule for numerical quadrature from univariate nodes and weights}
\usage{
quadratureProductRule(dim, nodes, weights)
}
\arguments{
\item{dim}{An integer that defines the dimension of the output quadrature formula}

\item{nodes}{A vector with univariate node points}

\item{weights}{A vector with univariate weights}
}
\value{
A list with a matrix of multivariate node points (each row is one point) and a vector of corresponding weights
}
\description{
Product rule for numerical quadrature from univariate nodes and weights
}
\examples{
require(statmod)
N <- 4; kind <- "hermite"
gauss <- gauss.quad(N,kind=kind)
w <- gauss$weights
x <- gauss$nodes
multQuad <- quadratureProductRule(2,x,w)
}
