% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGeneric.R, R/Pn_lognormalDensity.R,
%   R/Pn_logtDensity.R, R/Rn_Gauss.R, R/Rn_floorNorm.R, R/Rn_normalDensity.R,
%   R/Rn_tDensity.R, R/standardSimplex_Dirichlet.R, R/standardSimplex_exp_sum.R,
%   R/unitBall_normGauss.R, R/unitBall_polynomial.R, R/unitCube_BFN4.R,
%   R/unitCube_cos2.R, R/unitCube_floor.R, R/unitCube_max.R,
%   R/unitSphere_innerProduct1.R, R/unitSphere_polynomial.R
\name{evaluate}
\alias{evaluate}
\alias{evaluate,Pn_lognormalDensity,matrix-method}
\alias{evaluate,Pn_logtDensity,ANY-method}
\alias{evaluate,Rn_Gauss,matrix-method}
\alias{evaluate,Rn_floorNorm,matrix-method}
\alias{evaluate,Rn_normalDensity,matrix-method}
\alias{evaluate,Rn_tDensity,ANY-method}
\alias{evaluate,standardSimplex_Dirichlet,matrix-method}
\alias{evaluate,standardSimplex_exp_sum,matrix-method}
\alias{evaluate,unitBall_normGauss,matrix-method}
\alias{evaluate,unitBall_polynomial,matrix-method}
\alias{evaluate,unitCube_BFN4,matrix-method}
\alias{evaluate,unitCube_cos2,matrix-method}
\alias{evaluate,unitCube_floor,matrix-method}
\alias{evaluate,unitCube_max,matrix-method}
\alias{evaluate,unitSphere_innerProduct1,matrix-method}
\alias{evaluate,unitSphere_polynomial,matrix-method}
\title{Evaluate test function instance for a set of node points}
\usage{
evaluate(object, x)

\S4method{evaluate}{Pn_lognormalDensity,matrix}(object, x)

\S4method{evaluate}{Pn_logtDensity,ANY}(object, x)

\S4method{evaluate}{Rn_Gauss,matrix}(object, x)

\S4method{evaluate}{Rn_floorNorm,matrix}(object, x)

\S4method{evaluate}{Rn_normalDensity,matrix}(object, x)

\S4method{evaluate}{Rn_tDensity,ANY}(object, x)

\S4method{evaluate}{standardSimplex_Dirichlet,matrix}(object, x)

\S4method{evaluate}{standardSimplex_exp_sum,matrix}(object, x)

\S4method{evaluate}{unitBall_normGauss,matrix}(object, x)

\S4method{evaluate}{unitBall_polynomial,matrix}(object, x)

\S4method{evaluate}{unitCube_BFN4,matrix}(object, x)

\S4method{evaluate}{unitCube_cos2,matrix}(object, x)

\S4method{evaluate}{unitCube_floor,matrix}(object, x)

\S4method{evaluate}{unitCube_max,matrix}(object, x)

\S4method{evaluate}{unitSphere_innerProduct1,matrix}(object, x)

\S4method{evaluate}{unitSphere_polynomial,matrix}(object, x)
}
\arguments{
\item{object}{Test function that gets evaluated}

\item{x}{Matrix where each row represents one evaluation point}
}
\value{
Vector where each element is an evaluation of the test function for a node point (row in x)
}
\description{
evaluate is a generic function that evaluates the test function instance for a collection of evaluation points represented by a matrix. Each row is one evaluation point.
}
\author{
Klaus Herrmann
}
