% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGeneric.R, R/Rn_Gauss.R, R/Rn_floorNorm.R,
%   R/standardSimplex_Dirichlet.R, R/standardSimplex_exp_sum.R,
%   R/unitBall_normGauss.R, R/unitBall_polynomial.R, R/unitCube_cos2.R,
%   R/unitCube_floor.R, R/unitSphere_innerProduct1.R, R/unitSphere_polynomial.R
\name{getTags}
\alias{getTags}
\alias{getTags,Rn_Gauss-method}
\alias{getTags,Rn_floorNorm-method}
\alias{getTags,standardSimplex_Dirichlet-method}
\alias{getTags,standardSimplex_exp_sum-method}
\alias{getTags,unitBall_normGauss-method}
\alias{getTags,unitBall_polynomial-method}
\alias{getTags,unitCube_cos2-method}
\alias{getTags,unitCube_floor-method}
\alias{getTags,unitSphere_innerProduct1-method}
\alias{getTags,unitSphere_polynomial-method}
\title{Get tags for test function instance}
\usage{
getTags(object)

\S4method{getTags}{Rn_Gauss}(object)

\S4method{getTags}{Rn_floorNorm}(object)

\S4method{getTags}{standardSimplex_Dirichlet}(object)

\S4method{getTags}{standardSimplex_exp_sum}(object)

\S4method{getTags}{unitBall_normGauss}(object)

\S4method{getTags}{unitBall_polynomial}(object)

\S4method{getTags}{unitCube_cos2}(object)

\S4method{getTags}{unitCube_floor}(object)

\S4method{getTags}{unitSphere_innerProduct1}(object)

\S4method{getTags}{unitSphere_polynomial}(object)
}
\arguments{
\item{object}{Test function for which the tags are returned}
}
\value{
Vector with tags related to the function
}
\description{
getTags is a generic function that returns a vector of tags associated to the test function instance.
}
\author{
Klaus Herrmann
}
