% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{pIntRule}
\alias{pIntRule}
\title{Product rule for numerical quadrature from univariate nodes and weights}
\usage{
pIntRule(x, dim = NULL)
}
\arguments{
\item{x}{Either a list with two elements $nodes and $weights representing a one dimensionsal quadrature formula which are used for all dimensions, or a list where each element is a itself a list with two elements $nodes and $weights. Then for each dimension the respective quadrature rule is used.}

\item{dim}{An integer that defines the dimension of the output quadrature formula. Default is NULL. If dim is NULL then x has to be a list of quadrature rules (list of lists) and the dimensions is automatically generated.}
}
\value{
A list with a matrix of multivariate node points (each row is one point) and a vector of corresponding weights
}
\description{
Product rule for numerical quadrature from univariate nodes and weights
}
\examples{
require(statmod)
herm <- gauss.quad(2,"hermite")
lag <- gauss.quad(3,"laguerre")
qRule1 <- pIntRule(herm,2)
qRule2 <- pIntRule(list(herm,lag))
}
